Factored math and geometry into separate file
