+++ smooth_mandel.cpp
@@ -1,114 +1,22 @@
-#include <GL/glut.h>
 #include <complex>
 #include <iostream>
 #include <thread>
-#include "BlockingQueue.h"
 #include <cmath>
+#include <GL/glut.h>
+#include "mandelbrot.h"
+#include "BlockingQueue.h"
 
 using namespace std;
 
-namespace {
-double real_center = -0.85;
-double imag_center = 0.0;
-double width = 2.8;
-constexpr GLsizei window_width = 800;
-constexpr GLsizei window_height = 800;
-
-using GlutWindow = int;
-GlutWindow main_window;
-GlutWindow iterates_window;
-}
-
-class ComplexIterate
-{
-public:
-    typedef complex<double> ValueType;
-
-public:
-    ComplexIterate() = default;
-    ComplexIterate(const ComplexIterate & c) = default;
-
-    ComplexIterate(double r, double i)
-    : _start(r, i), _value(_start), _slow_value(_start)
-    { }
-    
-    void iterate()
-    {
-        if (!_escaped && !_bounded) {
-            _value = _value * _value + _start;
-            auto abs_sqr = _value.real()*_value.real() +
-            _value.imag()*_value.imag();
-            if (abs_sqr > escape_value) {
-                _escaped = true;
-                _adjusted_count = _count -
-                  log(log(abs_sqr) / log(escape_value))/log(2.0);
-            } else {
-                const double epsilon = 0.00000000000001;
-                const double real_diff = _value.real() - _slow_value.real();
-                if (real_diff < epsilon && real_diff > -epsilon) {
-                    const double imag_diff = _value.imag() - _slow_value.imag();
-                    if (imag_diff < epsilon && imag_diff > -epsilon) {
-                        _bounded = true;
-                    }
-                }
-            }
-
-            if (_count > 0 && _count % 2 == 0) {
-                _slow_value = _slow_value * _slow_value + _start;
-            }
-            ++_count;
-        }
-    }
-
-    ValueType getValue() const
-    {
-        return _value;
-    }
-
-    float getCount() const
-    {
-        if (_escaped) {
-            return _adjusted_count;
-        } else {
-            return _count;
-        }
-    }
-
-    bool escaped() const
-    {
-        return _escaped;
-    }
-
-    bool bounded() const
-    {
-        return _bounded;
-    }
-
-    static double escape_value;
-
-private:
-    ValueType _start;
-    ValueType _value;
-    ValueType _slow_value;
-    bool _escaped = false;
-    bool _bounded = false;
-    unsigned _count = 0;
-    double _adjusted_count;
-};
-
-namespace {
-// Most iterates to see in the iterate window
 struct IterateWindowData {
+    // Maximum iterates to see in the iterate window
     static constexpr std::size_t iterate_limit = 5000;
     std::size_t iterate_count = 0;
     double min_real;
     double max_real;
     double min_imag;
     double max_imag;
-} iterate_window_data;
-
-ComplexIterate iterates[IterateWindowData::iterate_limit];
-}
+};
 
 // `iterates` must be allocated, but this function will use placement new on
 // its elemenst.
@@ -171,126 +79,24 @@ IterateWindowData calculateIterates(ComplexIterate iterates[], double x, double 
     return result;
 }
 
-double ComplexIterate::escape_value = 10e100;
-
+// Application state
 namespace {
-constexpr unsigned subsample_width = 2;
-constexpr unsigned subsamples = 2 * subsample_width * subsample_width;
-}
+using GlutWindow = int;
+GlutWindow main_window;
+GlutWindow iterates_window;
 
-class Pixel
-{
-public:
-    typedef void (*ColorMapFunc)(bool, double, float &, float &, float &);
+double real_center = -0.85;
+double imag_center = 0.0;
+double real_width = 2.8;
 
-public:
-    Pixel() { }
+constexpr GLsizei window_width = 800;
+constexpr GLsizei window_height = 800;
 
-    Pixel(double left, double top, double width)
-    : _final(false), _left(left), _top(top), _width(width)
-    {
-        // Setup the sub-iterates.
-        double sub_width = _width / static_cast<double>(subsample_width);
-        double quarter_width = sub_width / 4.0;
-        double three_width = sub_width / 2.0 + quarter_width;
-        double x, y;
-        unsigned i, k;
-        unsigned iter = 0;
-        for (x = left + quarter_width, i = 0; i < subsample_width; x += sub_width, ++i) {
-            for (y = top - quarter_width, k = 0; k < subsample_width; y -= sub_width, ++k) {
-                new (&_sub_iterates[iter]) ComplexIterate(x, y);
-                ++iter;
-            }
-        }
-        for (x = left + three_width, i = 0; i < subsample_width; x += sub_width, ++i) {
-            for (y = top - three_width, k = 0; k < subsample_width; y -= sub_width, ++k) {
-                new (&_sub_iterates[iter]) ComplexIterate(x, y);
-                ++iter;
-            }
-        }
-    }
+// Iterates view globals:
+IterateWindowData iterate_window_data;
+ComplexIterate iterates[IterateWindowData::iterate_limit];
 
-    void iterate()
-    {
-        if (_final) {
-            return;
-        }
-
-        bool any_live = false;
-        for (unsigned i = 0; i < subsamples; ++i) {
-            _sub_iterates[i].iterate();
-            if (!_sub_iterates[i].escaped() && !_sub_iterates[i].bounded()) {
-                any_live = true;
-            }
-        }
-        computeColor(_red, _green, _blue);
-        if (!any_live) {
-            _final = true;
-        }
-    }
-
-    bool isFinal() const
-    {
-        return _final;
-    }
-
-    void color(unsigned char & r, unsigned char & g, unsigned char & b)
-    {
-        r = _red; g = _green; b = _blue;
-    }
-
-    void computeColor(unsigned char & r, unsigned char & g, unsigned char & b)
-    {
-        float r_sum = 0.0, g_sum = 0.0, b_sum = 0.0;
-        for (unsigned i = 0; i < subsamples; ++i) {
-            float red, green, blue;
-            colorMap(_sub_iterates[i].escaped(), _sub_iterates[i].getCount(), red, green, blue);
-            r_sum += red;
-            g_sum += green;
-            b_sum += blue;
-        }
-        r = r_sum * (255.0 / static_cast<double>(subsamples));
-        g = g_sum * (255.0 / static_cast<double>(subsamples));
-        b = b_sum * (255.0 / static_cast<double>(subsamples));
-    }
-
-    static ColorMapFunc colorMap;
-
-private:
-    bool _final;
-    // Subsampling for smoothness.
-    ComplexIterate _sub_iterates[32];
-    double _left, _top, _width;
-    unsigned char _red, _green, _blue;
-};
-
-
-void colorMap1(bool esc, double iter, float & r, float & g, float & b)
-{
-    if (!esc) {
-        r = 0.0; g = 0.271; b = 0.361;
-        return;
-    }
-
-    const float r_start = 1.0, g_start = 0.4, b_start = 0.2;
-    const float r_end = 0.0, g_end = 0.541, b_end = 0.722;
-    
-    const unsigned range = 100;
-
-    if (fmod((iter / range), 2) < 1) {
-        iter = fmod(iter, range);
-    } else {
-        iter = range - fmod(iter, range);
-    }
-
-    auto alpha = iter / static_cast<float>(range);
-
-    r = (1.0 - alpha) * r_start + alpha * r_end;
-    g = (1.0 - alpha) * g_start + alpha * g_end;
-    b = (1.0 - alpha) * b_start + alpha * b_end;
-}
-
-namespace {
+// Main window related globals:
 constexpr unsigned bin_width = 4;
 
 bool bin_finished[window_height / bin_width][window_width / bin_width];
@@ -339,6 +145,7 @@ void doBin()
 
 void idleFunc()
 {
+    // TODO: Use a thread pool. This is needlessly expensive.
     thread threads[thread_count];
 
     for (unsigned y = 0; y < window_height / bin_width; ++y) {
@@ -423,8 +230,6 @@ void renderScene()
     glFlush();
 }
 
-Pixel::ColorMapFunc Pixel::colorMap = colorMap1;
-
 void initialize(double real_center, double imag_center, double width)
 {
     cout << "real: " << real_center << " imag: " << imag_center
@@ -460,8 +265,8 @@ void mouseHandler(int button, int state, int x, int y)
               ((double)window_width / 2.0);
             double screen_y = (double)((window_height / 2) - y) /
               ((double)window_height / 2.0);
-            double new_x = screen_x * (width / 2.0) + real_center;
-            double new_y = screen_y * (width / 2.0) + imag_center;
+            double new_x = screen_x * (real_width / 2.0) + real_center;
+            double new_y = screen_y * (real_width / 2.0) + imag_center;
             iterate_window_data = calculateIterates(iterates, new_x, new_y);
             glutPostWindowRedisplay(iterates_window);
         }
@@ -482,14 +287,14 @@ void mouseHandler(int button, int state, int x, int y)
         double screen_y = (double)((window_height / 2) - y) /
           ((double)window_height / 2.0);
 
-        double new_x = screen_x * (width / 2.0) + real_center;
-        double new_y = screen_y * (width / 2.0) + imag_center;
-        double new_width = width * factor;
+        double new_x = screen_x * (real_width / 2.0) + real_center;
+        double new_y = screen_y * (real_width / 2.0) + imag_center;
+        double new_width = real_width * factor;
 
         initialize(new_x, new_y, new_width);
         real_center = new_x;
         imag_center = new_y;
-        width = new_width;
+        real_width = new_width;
     }
 }
 
@@ -502,7 +307,7 @@ int main(int argc, char * argv[])
     glutInitWindowSize(window_width, window_height);
     main_window = glutCreateWindow("Gradual Mandelbrot Rendering");
     
-    initialize(real_center, imag_center, width);
+    initialize(real_center, imag_center, real_width);
 
     glEnable(GL_TEXTURE_2D);
     glGenTextures(1, &mandel_texture);
